requirement.txt
blinker==1.7.0
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Flask==3.0.0
fonttools==4.47.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.3
opencv-python==4.9.0.80
packaging==23.2
pdf2image==1.17.0
pillow==10.2.0
pyocr==0.8.5
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
rapidfuzz==3.6.1
six==1.16.0
thefuzz==0.20.0
Werkzeug==3.0.1
